spotify_ex v0.1.0 Spotify.Cookies

Convinience setters and getters for auth cookies

Functions

get_access_cookie(conn)

Gets the access token

get_cookies_from_response(map)

Returns a map of token cookies from a parsed response body

get_refresh_cookie(conn)

Gets the access token

set_access_cookie(conn, string)

Sets the access token

set_cookies(conn, map)

Sets the refresh token and access token and returns the conn.

##Example:

Spotify.Cookies.set_cookies(conn, %{refresh_token: rt, access_token: at})

set_refresh_cookie(conn, string)

Sets the refresh token