spotify_ex v0.1.0 Spotify.Playlist

Playlist API endpoints

API calls in this module require valid Authorization headers. See the OAuth section for more details.

https://developer.spotify.com/web-api/playlist-endpoints/

Functions

by_category(id)

Get a category’s playlists. Spotify Documentation

iex> Spotify.Playlist.by_category("123")
"https://api.spotify.com/v1/browse/categories/123/playlists"

by_category(id, params)

Get a category’s playlists with additional parameters. Spotify Documentation

Valid params: country, limit, offset

Method: GET

Example:

iex> Spotify.Playlist.by_category("123", [country: "US", limit: 5])
"https://api.spotify.com/v1/browse/categories/123/playlists?country=US&limit=5"

featured()

Get a list of featured playlists. Spotify Documenation

Method: GET

iex> Spotify.Playlist.featured
"https://api.spotify.com/v1/browse/featured-playlists"

featured(params)

Get a list of featured playlists. Spotify Documenation

Valid params: locale, country, timestamp, limit, offset

Method: GET

iex> Spotify.Playlist.featured(country: "US")
"https://api.spotify.com/v1/browse/featured-playlists?country=US"

follow_playlist(owner_id, playlist_id)

Add the current user as a follower of a playlist. Spotify Documentation

Optional Body Params: public

Method: PUT

iex> Spotify.Playlist.follow_playlist("123", "456")
"https://api.spotify.com/v1/users/123/playlists/456/followers"

get_playlist(user_id, playlist_id)

Get a playlist owned by a Spotify user. Spotify Documentation

Method: GET

iex> Spotify.Playlist.get_playlist("123", "456")
"https://api.spotify.com/v1/users/123/playlists/456"

get_playlist(user_id, playlist_id, params)

Get a playlist owned by a Spotify user. Spotify Documentation

Method: GET ** Optional Params fields, market

iex> Spotify.Playlist.get_playlist("123", "456", market: "foo")
"https://api.spotify.com/v1/users/123/playlists/456?market=foo"

get_users_playlists(user_id)

Get a list of the playlists owned or followed by a Spotify user. Spotify Documentation

Method: GET

iex> Spotify.Playlist.get_users_playlists("123")
"https://api.spotify.com/v1/users/123/playlists"

get_users_playlists(user_id, params)

Get a list of the playlists owned or followed by a Spotify user. Spotify Documentation

Method: GET ** Optional Params: limit, offset

iex> Spotify.Playlist.get_users_playlists("123", limit: 5)
"https://api.spotify.com/v1/users/123/playlists?limit=5"

search(params)

Spotify Documentation

Method: GET

** Required Params: q

** Optional Params: limit, offset, market

iex> Spotify.Playlist.search(q: "foo", limit: 5)
"https://api.spotify.com/v1/search?type=playlist&q=foo&limit=5"

unfollow_playlist(owner_id, playlist_id)

Remove the current user as a follower of a playlist. Spotify Documentation

Method: DELETE

iex> Spotify.Playlist.follow_playlist("123", "456")
"https://api.spotify.com/v1/users/123/playlists/456/followers"