spotify_ex v0.0.3 Spotify.Authentication

Authenticates the Spotify user.

After your app is authorized, the user must be authenticated. A redirect URL is specified in the config folder. This is the URL that Spotify redirects to after authorization, and should ultimately end up hitting this module’s authenticate function. If the authorization is successful, the param code will be present.

If a refresh token still exists, the client will refresh the access token.

Functions

authenticate(conn, map)

The authorization code must be present from spotify or an exception will be raised. The token will be refreshed if possible, otherwise the app will request new access and request tokens.

## Example: ##

iex> Spotify.authenticate(conn, %{"code" => code})
iex> # {:ok, "access_token", conn}

iex> Spotify.authenticate(conn, %{"not_a_code" => invalid})
iex> AuthenticationError, "No code provided by Spotify. Authorize your app again"

tokens_present?(conn)

Checks for refresh and access tokens

## Example: ##

defmodule PlayListController do
  plug :check_tokens

  def check_tokens do
    unless Spotify.Authentication.tokens_present?(conn) do
      redirect conn, to: authorization_path(:authorize)
    end
  end
end